// src/types/environment.type.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-shadow */
export enum OS {
    WINDOWS = 'Windows',
    MAC = 'Mac',
    UBUNTU = 'Ubuntu',
    UNKNOWN = 'Unknown',
}

export enum BROWSER {
    CHROME = 'Chrome',
    SAFARI = 'Safari',
    FIREFOX = 'Firefox',
    POSTMAN = 'Postman',
    UNKNOWN = 'Unknown'
}